improve purge

+++ src/Keboola/Console/Command/DeletedProjectsPurge.php
@@ -1,7 +1,9 @@
 <?php
+
 namespace Keboola\Console\Command;
 
 use Keboola\ManageApi\Client;
+use Keboola\ManageApi\ClientException;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -16,65 +18,47 @@ class DeletedProjectsPurge extends Command
         $this
             ->setName('storage:deleted-projects-purge')
             ->setDescription('Purge deleted projects.')
+            ->addArgument('url', InputArgument::REQUIRED, 'URL of stack including https://')
             ->addArgument('token', InputArgument::REQUIRED, 'manage api token')
+            ->addArgument('projectIds', InputArgument::IS_ARRAY | InputArgument::REQUIRED, 'IDs of projects to purge (separate multiple IDs with a space)')
             ->addOption('ignore-backend-errors', null, InputOption::VALUE_NONE, "Ignore errors from backend and just delete buckets and workspaces metadata")
-        ;
+            ->addOption('force', null, InputOption::VALUE_NONE, 'Actually perform destructive operations (purge). Without this flag, the command will only simulate actions.');
     }
-
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
+        $url = $input->getArgument('url');
         $token = $input->getArgument('token');
         assert(is_string($token));
-
-        $fh = fopen('php://stdin', 'r');
-        if (!$fh) {
-            throw new \Exception('Error on input read');
-        }
-
+        $projectIds = $input->getArgument('projectIds');
         $ignoreBackendErrors = (bool) $input->getOption('ignore-backend-errors');
+        $force = (bool) $input->getOption('force');
 
         $output->writeln(sprintf(
             'Ignore backend errors %s',
             $ignoreBackendErrors ? 'On' : 'Off'
         ));
+        $output->writeln(sprintf(
+            'Force mode %s',
+            $force ? 'On (destructive operations will be performed)' : 'Off (no destructive operations will be performed)'
+        ));
 
         $client = new Client([
+            'url' => $url,
             'token' => $token,
         ]);
 
-        $lineNumber = 0;
-        while ($row = fgetcsv($fh)) {
-            if ($lineNumber === 0) {
-                $this->validateHeader($row);
-            } else {
-                $this->purgeProject(
-                    $client,
-                    $output,
-                    $ignoreBackendErrors,
-                    (int) $row[0],
-                    (string) $row[1]
-                );
-            }
-            $lineNumber++;
+        foreach ($projectIds as $projectId) {
+            $this->purgeProject(
+                $client,
+                $output,
+                $ignoreBackendErrors,
+                (int) $projectId,
+                $force,
+            );
         }
 
         return 0;
-    }
-
-    /**
-     * @param array<int, string|null> $header
-     */
-    private function validateHeader(array $header): void
-    {
-        $expectedHeader = ['id', 'name'];
-        if ($header !== $expectedHeader) {
-            throw new \Exception(sprintf(
-                'Invalid input header: %s Expected header: %s',
-                implode(',', $header),
-                implode(',', $expectedHeader)
-            ));
-        }
     }
 
     private function purgeProject(
@@ -82,12 +66,34 @@ class DeletedProjectsPurge extends Command
         OutputInterface $output,
         bool $ignoreBackendErrors,
         int $projectId,
-        string $projectName
+        bool $force,
     ): void {
+        try {
+            $deletedProject = $client->getDeletedProject($projectId);
+            if ($deletedProject['isPurged'] === true) {
+                $output->writeln(sprintf('<info>INFO</info> Project "%d" purged already.', $projectId));
+                return;
+            }
+        } catch (ClientException $e) {
+            if ($e->getCode() === 404) {
+                $output->writeln(sprintf('<error>Error</error>: Purge of the project "%d" not found.', $projectId));
+
+                return;
+            }
+            $output->writeln(sprintf('<error>Error</error>: Purge of the project "%d" is not possible due "%s".', $projectId, $e->getMessage()));
+            return;
+        }
+
+        $projectName = $deletedProject['name'] ?? 'unknown';
         $output->writeln(sprintf('Purge %s (%d)', $projectName, $projectId));
 
+        if (!$force) {
+            $output->writeln("[DRY-RUN] Would purge project $projectId");
+            return;
+        }
+
         $response = $client->purgeDeletedProject($projectId, [
-            'ignoreBackendErrors' => (bool) $ignoreBackendErrors,
+            'ignoreBackendErrors' => $ignoreBackendErrors,
         ]);
         $output->writeln(" - execution id {$response['commandExecutionId']}");
 
@@ -98,9 +104,16 @@ class DeletedProjectsPurge extends Command
             if (time() - $startTime > $maxWaitTimeSeconds) {
                 throw new \Exception("Project {$projectId} purge timeout.");
             }
-            sleep(1);
+            sleep(2);
+            $output->writeln(
+                sprintf(' - - Waiting for project "%s" (%s) to be purged: execution id %s',
+                    $projectName,
+                    $projectId,
+                    $response['commandExecutionId'],
+                )
+            );
         } while ($deletedProject['isPurged'] !== true);
 
-        $output->writeln(sprintf('Purge done %s (%d)', $projectName, $projectId));
+        $output->writeln(sprintf('<info>Purge done "%s" (%d)</info>', $projectName, $projectId));
     }
 }
